requirements.txt
gitpython
openai
pyttsx3
selenium
ffmpeg-python
python-dotenv
webdriver-manager
requests

